fix(ocean): handle failed deletions of launch specs
When a creation of a stack fails, CFN issues a corresponding delete operation to clean up the resources and roll-back (delete) the resources.

This adds proper handling for this scenario: If the corresponding delete operation failed because Spot API indicates that the launch-spec does not exist, we mark the deletion operation as successful - letting the stack be reverted/deleted without user-intervention.